Support named temporary credentials
Credentials are {id, key, cert}.  A "simple" temporary credential has
the issing client as its `id`, scopes and expiry in the cert, all signed
to generate a new key.  "Named" temporary credentials further include a
name and an issuer in the certificate; the issuer is the clientId that
created the credential (and the corresponding accessToken is used in the
signature), while the name is the clientId that should be supplied as
`id`.  The result is that temporary credentials can have a meaningful
id that is useful for logging and for identifying the credentials.