[refactor] update resolver generation and execution
The core of resolution was not and is not a DFA -- it's just a Trie.
That is now cleaned up and shipped off to its own module.  It's also
simplified, in that it no longer represents recursive expansion.  Very
soon (parameterized roles) such a thing will be impossible.

Resolver execution is also substantially refactored: instead of
searching for a fixed point, which soon will not exist, excution
involves a recursive expansion of roles.  The recursion is flattened
into a queue, with the scopes added to the queue becoming the result.
This is a different interface than before, so there is some
compatibility in place that will be removed later with a refactor of
scoperesolver.js.  That will also entail a move of the remainder of
dfa.js into scoperesolver.js.

This new approach changes performance drastically: initial generation of
a resolver is quite a bit faster, but execution is 10-100x slower.  The
worst cases are those that involve recursing over just about every role,
only to normalize to a very short list of scopes.

The trie execution return value is a little unusual, but the purpose
will become clear soon enough.

This also consolidates the various tests for scoperesolver.js.