Prohibit parameterized roles with cycles
Cycles with parameterized roles can potentially cause infinite
expansion; for example

  a* -> assume:bc%
  b* -> assume:a%d

here ax -> bcx -> acxd -> bccxd -> ...

We therefore prohibit compilation of such sets of roles and also check
in REST API calls.

Note, however, that at this stage it is still possible to use the API to
write such roles in a race condition, resulting in a failure of the auth
service.  This will be addressed in the next commit.