Support parameterized roles (finally!)
This adds support for `<..>` in role expansions.  On execution, the
portion of the input matching a final `*` in a roleid -- or the empty
string if the roleId does not end in `*` -- substituted for `<..>`.
There is a special case that if the input ends with `*` then it consumes
the rest of the scope.  Examples, given role `assume:ab* -> de<..>f,
gh<..>`:

 * `assume:abc -> decf ghc`
 * `assume:abcd* -> decd*, ghcd*`
 * `assume:ab* -> de*, gh*`
 * `assume:a* -> de*, gh*`