Optimize resolver's seen tracking
Modify the resolver to keep a normalized `seen` scopeset and use the
work queue to flatten recursion only for scopes that are not already
satisfied by `seen`. This has a big impact on massive expansions that
end up reducing to `*` or `assume:*`: once the star scope is seen,
nothing else goes on the queue, so recursion peters out quickly.

The final result is the seen queue.

Times in my testing are down to <10ms for all of the "real roles"
performance tests. It's 500-1000x slower for some of the pathological
examples, which is about where the old solution was.