Add percent symbol to the client ID schema
Auth0 doesn't provide a nice user ID for GitHub handles. For GitHub we
receive IDs of the form github|12334. For better user IDs, we append the
GitHub nickname to the ID (e.g, github|1234/helfi92). Encoding the
latter gives github%7C1234%2Fhelfi92. To be able to use this, % needs to
be included in the Client ID schema.