Bug 1481944 - update to Hawk-4.1.2
Hawk does not follow semver, and does not have a changelog. The only
thing described as "breaking" is
hueniverse/hawk@f11626d1a651f505700d26af7bb14fabe7fdf65d. This is a
change to the nonceFunc API; auth does not use nonceFunc, but had some
unused support for it which is now removed. Attempts to use that support
will fail with an assertion error.