Bug 1508849 - implement **-scope protection differently
The old solution used a negative lookbehind, which per
https://json-schema.org/understanding-json-schema/reference/regular_expressions.html#regular-expressions
is not supported in json-schema.  In fact, it's not supported in even
fairly recent versions of Node.  This causes problems when validating
schemas using those old versions of node (or using json-schema
validators that do not support this form).

So, the new solution is to explicitly forbid those forms where used in
creating clients and roles.  The result is much the same, as evidenced
by the minimal changes to the tests.